requirements.txt
alembic==1.6.5
blinker==1.4
boto3==1.18.6
botocore==1.21.6
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
Flask==2.0.1
Flask-Admin==1.5.8
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
Mako==1.1.4
MarkupSafe==2.0.1
oauthlib==3.1.1
psycopg2-binary==2.9.1
python-dateutil==2.8.2
python-dotenv==0.18.0
python-editor==1.0.4
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
SQLAlchemy==1.4.22
urllib3==1.26.6
Werkzeug==2.0.1
WTForms==2.3.3
Flask-JWT-Extended==4.2.3